WIP on CheckinsUnassignedSubview. * Bunch of refactoring on the Sequelize models   to get one->many to work both ways. * TODO: assignsAssign blows up on the update still.

+++ src/models/AbstractModel.ts
@@ -31,7 +31,7 @@ export abstract class AbstractModel<Model> extends Model {
         autoIncrement: true,
         comment: "Primary key for this instance.",
         primaryKey: true,
-        type: DataType.BIGINT,
+        type: DataType.INTEGER,
     })
     id?: number;
